app/requirements.txt
databricks-sql-connector
databricks-sdk
psycopg2-binary
pandas
requests
